fix: align remaining client DTOs with spec (all low-severity audit items)
- Change retryAfterMs from Long to Integer across ReservationResult,
  DecisionResult, DryRunResult, CyclesProtocolException, and
  CyclesLifecycleService to match server's Integer type
- Add ReservationSummaryResult for list endpoint (spec's ReservationSummary
  vs ReservationDetail distinction); update ReservationListResult to use it
- Fix Caps.toolAllowlist/toolDenylist defaults from List.of() to null when
  absent, matching spec semantics (absence vs empty array)
- Add typed status enums to CommitResult (ReservationStatus.COMMITTED),
  ReleaseResult (ReservationStatus.RELEASED), ExtendResult
  (ReservationStatus.ACTIVE), and EventResult (new EventStatus.APPLIED)
- Add EventStatus enum for event-specific APPLIED status value

+++ cycles-client-java-spring/src/main/java/io/runcycles/client/java/spring/context/CyclesLifecycleService.java
@@ -99,7 +99,7 @@ public class CyclesLifecycleService {
         List<String> affectedScopes = resResult.getAffectedScopes();
         String scopePath = resResult.getScopePath();
         Amount reserved = resResult.getReserved();
-        Long retryAfterMs = resResult.getRetryAfterMs();
+        Integer retryAfterMs = resResult.getRetryAfterMs();
         List<Balance> balances = resResult.getBalances();
 
         // Validate decision field
