+++ cycles-client-java-spring/src/main/java/io/runcycles/client/java/spring/model/ReservationResult.java
@@ -16,13 +16,13 @@ public class ReservationResult {
     private final Amount reserved;
     private final Caps caps;
     private final String reasonCode;
-    private final Long retryAfterMs;
+    private final Integer retryAfterMs;
     private final List<Balance> balances;
 
     private ReservationResult(Decision decision, String reservationId,
                               List<String> affectedScopes, Long expiresAtMs,
                               String scopePath, Amount reserved, Caps caps,
-                              String reasonCode, Long retryAfterMs,
+                              String reasonCode, Integer retryAfterMs,
                               List<Balance> balances) {
         this.decision = decision;
         this.reservationId = reservationId;
@@ -48,7 +48,7 @@ public class ReservationResult {
                 map.get("reserved") instanceof Map<?, ?> m ? Amount.fromMap((Map<String, Object>) m) : null,
                 Caps.fromMap(map.get("caps") instanceof Map<?, ?> m ? (Map<String, Object>) m : null),
                 map.get("reason_code") instanceof String s ? s : null,
-                map.get("retry_after_ms") instanceof Number n ? n.longValue() : null,
+                map.get("retry_after_ms") instanceof Number n ? n.intValue() : null,
                 Balance.listFromRaw(map.get("balances") instanceof List<?> l ? l : null)
         );
     }
@@ -61,7 +61,7 @@ public class ReservationResult {
     public Amount getReserved() { return reserved; }
     public Caps getCaps() { return caps; }
     public String getReasonCode() { return reasonCode; }
-    public Long getRetryAfterMs() { return retryAfterMs; }
+    public Integer getRetryAfterMs() { return retryAfterMs; }
     public List<Balance> getBalances() { return balances; }
 
     public boolean isAllowed() { return decision == Decision.ALLOW || decision == Decision.ALLOW_WITH_CAPS; }
